pyproject.toml
[tool.ruff]
line-length = 100

[tool.poetry.group.migration.dependencies]
alembic = "^1.16.4"


[tool.poetry.group.dev.dependencies]
boto3-stubs = {extras = ["s3"], version = "^1.40.21"}
pyright = "^1.1.406"
types-pyjwt = "^1.7.1"

[project]
name = "velvet"
version = "0.1.0"
description = "Бэкенд для магазина тканей"
authors = [
    {name = "ma4etea",email = "alexvlod85@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <3.12"
dependencies = [
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (==4.0.1)",
    "passlib (>=1.7.4,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "black (>=25.1.0,<26.0.0)",
    "ruff (>=0.12.5,<0.13.0)",
    "backoff (>=2.2.1,<3.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
    "inflection (>=0.5.1,<0.6.0)",
    "celery (>=5.5.3,<6.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "psycopg (>=3.2.9,<4.0.0)",
    "aioboto3 (>=15.1.0,<16.0.0)",
    "urllib3 (>=2.5.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
